[ALPS02213688] Display:Kernel-3.18 Driver Porting
[Detail]
fix build error
[Solution]
fix build error
[Feature] Display Driver (DSI Interface)

MTK-Commit-Id: fe2d89b57400f276871012e9ebb4087ecdda5db5

CR-Id: ALPS02213688

+++ drivers/misc/mediatek/video/mt6755/dispsys/ddp_reg.h
@@ -1134,8 +1134,8 @@ static inline unsigned long disp_addr_convert(unsigned long va)
 		if (handle == NULL) { \
 			mt_reg_sync_writel(val, (volatile unsigned long*)(reg32));\
 		} else { \
-			dprec_reg_op((void *)handle, reg32, val, 0x00000000);\
-			cmdqRecWrite((cmdqRecHandle)handle, disp_addr_convert(reg32), val, ~0); \
+			dprec_reg_op((void *)handle, (unsigned long)(reg32), val, 0x00000000);\
+			cmdqRecWrite((cmdqRecHandle)handle, disp_addr_convert((unsigned long)reg32), val, ~0); \
 		}  \
 	} while (0)
 
@@ -1159,7 +1159,7 @@ static inline unsigned long disp_addr_convert(unsigned long va)
 			while (!((DISP_REG_GET(reg32))&val))\
 				; \
 		} else { \
-			dprec_reg_op(handle, (unsigned int)(reg32), val, 0x00000000);\
+			dprec_reg_op(handle, (unsigned long)(reg32), val, 0x00000000);\
 			cmdqRecPoll(handle, disp_addr_convert((unsigned long)(reg32)), val, mask); \
 		}  \
 	} while (0)
